implement appbar with builder pattern
